2.9 main ( triangulos )

+++ scr/es/uah/matcomp/mp/e1/ejerciciosclases/P2/e2_9/MyPoint.java
@@ -1,8 +1,8 @@
 package es.uah.matcomp.mp.e1.ejerciciosclases.P2.e2_9;
 
 public class MyPoint {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     ////////////////////////////////////////////////////////////////
     public MyPoint() {
@@ -10,35 +10,35 @@ public class MyPoint {
         this.y = 0;
     }
 
-    public MyPoint(int x, int y) {
+    public MyPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
     ////////////////////////////////////////////////////////////////
 
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int[] getXY() {
-        int[] ints = new int[]{x,y};
+    public double[] getXY() {
+        double[] ints = new double[]{x,y};
         return ints;
     }
 
-    public void setXY(int x, int y){
+    public void setXY(double x, double y){
         this.x = x;
         this.y = y;
     }
@@ -47,15 +47,15 @@ public class MyPoint {
         return String.format("(%s,%s)",x,y);
     }
 
-    public double distance(int x, int y) {
-        int deltaX = this.x - x;
-        int deltaY = this.y - y;
+    public double distance(double x, double y) {
+        double deltaX = this.x - x;
+        double deltaY = this.y - y;
         return Math.sqrt(deltaX*deltaX + deltaY*deltaY);
     }
 
     public double distance(MyPoint another) {
-        int deltaX = this.x - another.x;
-        int deltaY = this.y - another.y;
+        double deltaX = this.x - another.x;
+        double deltaY = this.y - another.y;
         return Math.sqrt(deltaX*deltaX + deltaY*deltaY);
     }
